separate out search result into its own file
